requirement.txt
amqp==2.3.1
backcall==0.1.0
billiard==3.5.0.3
celery==4.1.1
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
Django==2.0.5
django-celery-results==1.0.1
django-cors-headers==2.2.0
django-mysql==2.2.2
djangorestframework==3.8.2
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
kombu==4.2.0
mysqlclient==1.3.12
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.3
python-quickbooks==0.7.4
pytz==2018.4
rauth==0.7.3
redis==2.10.6
requests==2.18.4
simplegeneric==0.8.1
simplejson==3.15.0
six==1.11.0
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
